The server is up and running but that's about it
I have created the server and I have also added the express.js on top of node in order to use routing for different Get methods. On the frontend I basically have an index.js which handles the fetch method, where I use template literals to bind the fetched data to a newly created div.